+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `showMigration`
 * Extend allowable version bounds for aeson and splitmix
+* Add support for postgres' oid column type
 
 ## 0.1.2.0